Move patching react internals to a separate file
